feat(lib): properly handle basename matching
This commit limits glob matching similar to lint-staged. Basically,
globs only happen against relative paths. The basename (or matchBase)
option is set to true if the pattern does contain a path separator. So,
*.js can match foo/sample.js, or sample.js, or foo/bar/sample.js by
default. If the glob is something like foo/*.js, then basename is set to
false for default expected globbing.

The next step is to support a "relative" option to convert absolute
paths to relative paths for command args. Then I think the behavior is
consistent with lint-staged.